b - networking fixes for HPUX, fixes pbs_iff (PACCAR)
The socklen_t test in configure was enhanced with a special case with
gcc+xnet on HPUX.  Turns out HPUX's gcc defines socklen_t by default,
but the default networking stack uses int and breaks at run-time.  The
solution is to link with libxnet.  Because socken_t was defined in a
case where we don't actually want to use socklen_t, al usage of
socklen_t is replaced with torque_socklen_t.

Also, HPUX doesn't define ntohl() and friends in the proper headers,
so that was fixed with further configure tests.

No more socklen_t!  torque_socklen_t!  Hopefully this is the final
chapter of TORQUE's long tortured history of socklen_t problems.



git-svn-id: svn://opensvn.adaptivecomputing.com/torque/trunk@930 8f940c70-5916-0410-ac46-d1fa2fa6ea29