added "rerunable" to types of jobs allowed in queue disallowed_types acl
pbs_server now keeps a list of array_job_list structs, one for each job array
this struct contains a tlist_head for a list of all jobs that belong to that particular array.  Each job also has a pointer to this struct, and a list_link so it is linked into the list.


git-svn-id: svn://opensvn.adaptivecomputing.com/torque/trunk@1372 8f940c70-5916-0410-ac46-d1fa2fa6ea29