This change takes care of a problem in large clusters where an outgoing tcp buffer is not large enough to hold the entire job spec to start a job. Previous to this fix the largest a tcp buffer could be was 262144 bytes as defined by THE_BUF_SIZE in dis.h.
In order to allow a dynamic allocation of buffer space for tcp it was
necessary to separate the dis functions into tcp and udp calls. The
existing dis calls are still in the code and for the most part are used
for udp packets. The several new dis functions were created from the old
dis functions and are now prefixed with a tcp_. For example
decode_DIS_Authen is now tcp_decode_DIS_Authen for tcp calls. Because
tcp uses the global array tcparry this is where the static buffer was
changed to a pointer with buffer space that could be expanded. The allocation
changes are mainly in DIS_tcp_setup and tcp_puts.



git-svn-id: svn://opensvn.adaptivecomputing.com/torque/branches/2.3-fixes@3082 8f940c70-5916-0410-ac46-d1fa2fa6ea29