First checkin for changes to TORQUE 3.0. Changes include the replacement of the tinsert, tfind, etc. functions with an AVL_ prefixed set of utilities that use and AVL balanced tree to sort known host addresses.
Also included is functionality that allows more than one mom to run on a node.
This is also known as the multi-mom.

To aid TORQUE in spanning clusters with 20,000 plus nodes (200,000 is the 
target) a job_radix option was added that will use a radix to connect sisters
in a multinode job. The Mother Superior no longer communicates directly
with all sisters in the job. Instead she communicates to the number of 
sisters in the job_radix who in turn communicate with the same number of 
sisters in the job.



git-svn-id: svn://opensvn.adaptivecomputing.com/torque/branches/3.0-alpha@3581 8f940c70-5916-0410-ac46-d1fa2fa6ea29